🚧 Wip: 작업중

+++ apps/web/src/components/contestform/molecule/ContestUploader.tsx
@@ -3,11 +3,12 @@
 import React, { useState, useRef } from "react";
 import Image from "next/image";
 import { UploadCloud, XCircleIcon } from "lucide-react";
-import type { CreateContestType } from "@/types/request/requestType";
+import type { CreateFeedType } from "@/types/request/requestType";
 import {
   deleteFileFromS3,
   uploadFileToS3,
 } from "@/actions/common/awsMediaUploader";
+import { MediaAssetConverter } from "@/functions/utils/mediaListConverter";
 import { UuidConverter } from "@/functions/utils/uuidConverter";
 
 export interface PreviewImageType {
@@ -18,11 +19,11 @@ export interface PreviewImageType {
 function ImageUploader({
   setPayload,
 }: {
-  setPayload: React.Dispatch<React.SetStateAction<CreateContestType>>;
+  setPayload: React.Dispatch<React.SetStateAction<CreateFeedType>>;
 }) {
   const mediaRef = useRef<HTMLInputElement>(null);
-  const [previewMedia, setPreviewMedia] = useState<PreviewImageType | null>(
-    null,
+  const [mediaList, setMediaList] = useState<PreviewImageType[]>(
+    [] as PreviewImageType[],
   );
 
   const handleFeedImage = async (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -31,38 +32,45 @@ function ImageUploader({
     if (!files || files.length === 0) return;
 
     const file = files[0];
-
+    const fileType = file.type.startsWith("image") ? "IMAGE" : "VIDEO";
     const fileExtention = file.name.split(".").pop();
     const uniqueFileName = `${UuidConverter()}.${fileExtention}`;
 
     const s3Url = await uploadFileToS3(file, uniqueFileName);
 
     if (!s3Url) return;
+    setMediaList((prev) => {
+      const updated = [...prev];
+      updated.push({ s3Url, fileType });
+      return updated;
+    });
 
-    setPreviewMedia({ s3Url, fileType: "IMAGE" });
+    const convertedMedia = await MediaAssetConverter(s3Url, fileType);
 
-    setPayload((prev) => ({
-      ...prev,
-      mediaUrl: s3Url,
-      mediaType: "IMAGE",
-    }));
+    setPayload((prev) => {
+      const updatedPayload = { ...prev };
+      updatedPayload.mediaList = [convertedMedia];
+      return updatedPayload;
+    });
   };
 
-  const handleDeleteImage = async (s3Url: string) => {
+  const handleDeleteImage = async (s3Url: string, fileType: string) => {
     try {
-      if (s3Url) {
+      if (s3Url && fileType) {
         const res = await deleteFileFromS3(s3Url);
         if (!res) throw new Error("Failed to delete from S3");
       }
 
-      // 삭제 후 초기화
-      setPreviewMedia(null);
-      setPayload((prev) => ({
-        ...prev,
-        mediaUrl: "",
-        mediaType: "",
-      }));
-
+      // 이미지 삭제 후 해당 슬롯 비우기
+      setPayload((prev) => {
+        const updatedPayload = { ...prev };
+        updatedPayload.mediaList = [];
+        return updatedPayload;
+      });
+      setMediaList((prev) => {
+        const updated = prev.filter((media) => media.s3Url !== s3Url);
+        return updated;
+      });
       // input file 초기화
       if (mediaRef.current) {
         mediaRef.current.value = "";
@@ -90,14 +98,18 @@ function ImageUploader({
         id="feedImg"
         name="feedImg"
         className="hidden"
-        accept="image/*"
+        accept="image/*,video/*"
+        multiple
         onChange={(e) => {
           void handleFeedImage(e);
         }}
       />
-      <div className="mt-4 flex justify-center">
-        {previewMedia ? (
-          <div className="relative w-[120px] h-[150px] bg-gray-200 rounded-lg flex items-center justify-center text-gray-400 border-dashed border-2 overflow-hidden">
+      <ul className="flex gap-4 mt-4 flex-wrap w-full justify-center">
+        {mediaList.map((previewMedia: PreviewImageType) => (
+          <li
+            key={previewMedia.s3Url}
+            className="relative w-[120px] h-[150px] bg-gray-200 rounded-lg flex items-center justify-center text-gray-400 border-dashed border-2 overflow-hidden"
+          >
             {previewMedia.fileType === "IMAGE" ? (
               <Image
                 src={previewMedia.s3Url}
@@ -106,17 +118,27 @@ function ImageUploader({
                 width={150}
                 height={120}
               />
-            ) : null}
+            ) : (
+              <video width="100%" preload="metadata" autoPlay muted>
+                <source src={previewMedia.s3Url} />
+                <track kind="captions" default />
+              </video>
+            )}
             <button
               type="button"
-              onClick={() => void handleDeleteImage(previewMedia.s3Url)}
+              onClick={() =>
+                void handleDeleteImage(
+                  previewMedia.s3Url,
+                  previewMedia.fileType,
+                )
+              } // fileType 전달
               className="absolute inset-0 flex items-center justify-center bg-black bg-opacity-50 text-white opacity-0 hover:opacity-100 transition-opacity"
             >
               <XCircleIcon />
             </button>
-          </div>
-        ) : null}
-      </div>
+          </li>
+        ))}
+      </ul>
     </div>
   );
 }
